pyproject.toml
[project]
name = "pixelsmith"
version = "0.1.0"
description = "Generate 8-bit pixel art from text prompts using SDXL + LoRA"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "torch>=2.1",
    "diffusers>=0.27",
    "transformers>=4.36",
    "accelerate>=0.25",
    "peft>=0.9",
    "safetensors>=0.4",
    "Pillow>=10.0",
    "numpy>=1.26",
    "platformdirs>=4.0",
]

[project.optional-dependencies]
mcp = ["fastmcp>=2.0,<4", "anyio>=4.0"]
dev = ["pytest>=9.0", "pytest-asyncio>=1.3", "ruff>=0.15", "pyright>=1.1"]

[project.scripts]
pixelsmith-mcp = "pixelsmith.mcp.server:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "standard"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["gpu: requires CUDA GPU"]
